requirements.txt
huggingface-hub
python-dotenv
streamlit
pillow
requests

